third_party/blink/web_tests/platform/generic/fast/events/pointerevents/touch-pointer-mouse-expected.txt

Verifies that canceling pointerdowns for touches correctly suppresses mouseevents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

=== With primary pointerdown uncanceled ===
-- Tap with paired GEs --
Received pointerdown primary
Received touchstart 
Received pointerup primary
Received touchend 
Received mousemove 
Received mousedown 
Received mouseup 
Received click 
-- Tap without paired GEs --
Received pointerdown primary
Received touchstart 
Received pointerup primary
Received touchend 
Received pointerdown primary
Received touchstart 
Received pointerup primary
Received touchend 
Received mousemove 
Received mousedown 
Received mouseup 
Received click 
Received mousemove 
Received mousedown 
Received mouseup 
Received click 
-- Long-press with paired GEs --
Received pointerdown primary
Received touchstart 
Received pointercancel primary
Received touchcancel 
Received mousemove 
-- Long-press without paired GEs --
Received pointerdown primary
Received touchstart 
Received pointercancel primary
Received touchcancel 
Received mousemove 
Received mousemove 


=== With primary pointerdown canceled ===
-- Tap with paired GEs --
Received pointerdown primary,canceled
Received touchstart 
Received pointerup primary
Received touchend 
Received click 
-- Tap without paired GEs --
Received pointerdown primary,canceled
Received touchstart 
Received pointerup primary
Received touchend 
Received pointerdown primary,canceled
Received touchstart 
Received pointerup primary
Received touchend 
Received click 
Received mousemove 
Received mousedown 
Received mouseup 
Received click 
-- Long-press with paired GEs --
Received pointerdown primary,canceled
Received touchstart 
Received pointercancel primary
Received touchcancel 
-- Long-press without paired GEs --
Received pointerdown primary,canceled
Received touchstart 
Received pointercancel primary
Received touchcancel 
Received mousemove 
Received mousemove 


=== With primary pointerdown uncanceled ===
-- Tap with paired GEs --
Received pointerdown primary
Received touchstart 
Received pointerup primary
Received touchend 
Received mousemove 
Received mousedown 
Received mouseup 
Received click 
-- Tap without paired GEs --
Received pointerdown primary
Received touchstart 
Received pointerup primary
Received touchend 
Received pointerdown primary
Received touchstart 
Received pointerup primary
Received touchend 
Received mousemove 
Received mousedown 
Received mouseup 
Received click 
Received mousemove 
Received mousedown 
Received mouseup 
Received click 
-- Long-press with paired GEs --
Received pointerdown primary
Received touchstart 
Received pointercancel primary
Received touchcancel 
Received mousemove 
-- Long-press without paired GEs --
Received pointerdown primary
Received touchstart 
Received pointercancel primary
Received touchcancel 
Received mousemove 
Received mousemove 


PASS successfullyParsed is true

TEST COMPLETE
